Update Sound Effect guide with more instructions
